requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain_community
openai
qdrant-client
edge-tts
tiktoken
langchain_openai